pyproject.toml
[project]
name = "dutyproject"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Lera Dzanagova", email = "lera.dzanagova@gmail.com"},
]
dependencies = [
    "asgiref==3.8.1",
    "Django==5.1.6",
    "django-environ==0.12.0",
    "django-filter==25.1",
    "django-simple-history==3.8.0",
    "djangorestframework==3.15.2",
    "psycopg2-binary==2.9.6",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.1",
    "six==1.17.0",
    "slack-sdk==3.35.0",
    "sqlparse==0.5.3",
    "tzdata==2025.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
